refactor(document-generation): add abstraction to avoid duplication of generateFiles method
