+++ apps/document-generation/src/documents/CreditReportDispute.js
@@ -1,18 +1,10 @@
 // @flow
 
+import Document from "./Document";
 import PDFEngine from "../engines/PDFEngine";
 
-class CreditReportDispute implements DocumentGenerator {
+class CreditReportDispute extends Document implements DocumentGenerator {
   engine = PDFEngine;
-  generateFiles = async (data: mixed) => {
-    const pathsToTemplate = this.templates.map(
-      template => `../templates/${template}`
-    );
-
-    const processedFiles = await this.engine.process(data, pathsToTemplate[0]);
-
-    return [processedFiles];
-  };
   slug = "credit-report-dispute";
   templates = [`${this.slug}/0.hbs`];
   version = "v1";
