+++ apps/document-generation/src/worker.js
@@ -1,6 +1,7 @@
 // @flow
 
 import Dispute from "./models/Dispute";
+import Document from "./documents/Document";
 import { findBySlug } from "./documents";
 import User from "./models/User";
 
@@ -13,8 +14,8 @@ const run = async ({ userId, disputeId }: Params) => {
   const user = await User.findById(userId);
   const dispute = await Dispute.findById(disputeId);
   const data = { dispute, user };
-  const Document = findBySlug(dispute.toolId);
-  const files = await Document.generateFiles(data);
+  const DocumentHandler: Document = findBySlug(dispute.toolId);
+  const files = await DocumentHandler.generateFiles(data);
 
   return files;
 };
